Add files via upload
The bundle of codes used for FY2021 central and eastern North America aftershock sequence study (USGS EHP External Grant G21AP10312).
Disclaimer: The views and conclusions contained in this document are those of the authors and should not be interpreted as representing the opinions or policies of the U.S. Geological Survey. Mention of trade names or commercial products does not constitute their endorsement by the U.S. Geological Survey.